Introduction
AI can create original works through machine learning and a variety of self-learning algorithms that extract information from data. The creation of AI art requires the input of both the human artist and the AI system; however, the degree of autonomy can vary widely, and the final product is highly dependent on the accuracy of the training data.